+++ packages/storybook/stories/constant/theme/highlight.ts
@@ -1,0 +1,8 @@
+import { HighlightCustomTheme, themes } from "@lmarcel/highlight";
+
+export const storybookTheme = HighlightCustomTheme.extends(themes.oneDark, {
+  numbersBorderColor: "#1ea7fd",
+  backgroundColor: "#272727",
+  numbersBackgroundColor: "#2b2a2a",
+  numbersColor: "#cfcfcf"
+});